发团管理

内外转
   报价
   游客小组
   中转安排
------------精度控制--------

See merge request !799

+++ js/template/arrange/transit/rule.js
@@ -65,7 +65,7 @@ define(function(require, exports) {
 						$ele: $that.find('input[name="busReduceMoney"]'),
 						rules: [
 							{
-								type: 'int',
+								type: 'float',
 								errMsg: '数据格式不正确'
 							}	
 						]
@@ -75,7 +75,7 @@ define(function(require, exports) {
 						$ele: $that.find('input[name="busPayedMoney"]'),
 						rules: [
 							{
-								type: 'int',
+								type: 'float',
 								errMsg: '数据格式不正确'
 							}	
 						]
@@ -112,7 +112,7 @@ define(function(require, exports) {
 						$ele: $that.find('input[name="hotelPrice"]'),
 						rules: [
 							{
-								type: 'int',
+								type: 'float',
 								errMsg: '单价格式不正确'
 							}	
 						]
@@ -136,7 +136,7 @@ define(function(require, exports) {
 						$ele: $that.find('input[name="hotelReduceMoney"]'),
 						rules: [
 							{
-								type: 'int',
+								type: 'float',
 								errMsg: '数据格式不正确'
 							}	
 						]
@@ -146,7 +146,7 @@ define(function(require, exports) {
 						$ele: $that.find('input[name="hotelPayedMoney"]'),
 						rules: [
 							{
-								type: 'int',
+								type: 'float',
 								errMsg: '数据格式不正确'
 							}	
 						]
